--- src/main/java/com/hhplus/cleanarchi/repository/CourseDetailRepository.java
@@ -1,10 +1,0 @@
-package com.hhplus.cleanarchi.repository;
-
-import com.hhplus.cleanarchi.entity.CourseDetail;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public interface CourseDetailRepository {
-    List<CourseDetail> findAvailableCoursesByDateAndCapacity(LocalDate date);
-}
